Fixed issue #249.

+++ src/bgfx.cpp
@@ -710,7 +710,7 @@ namespace bgfx
 		BX_WARN(invalidHandle != m_key.m_program, "Program with invalid handle");
 		if (invalidHandle != m_key.m_program)
 		{
-			m_key.m_depth  = _depth;
+			m_key.m_depth  = (uint32_t)_depth;
 			m_key.m_view   = _id;
 			m_key.m_seq    = s_ctx->m_seq[_id] & s_ctx->m_seqMask[_id];
 			s_ctx->m_seq[_id]++;
